make_lib_archive no longer appends a .zip extension, this allows other extensions for the archive as well.
Make make_lib_archive a method of the py2exe class, to allow
overriding in a subclass. It also allows the compression to be set -
default is still uncompressed.

The python dll cannot be copied in the same way as other dlls, it must
be in the directory of the executable(s) to be build.